+++ artifact.sh
@@ -2,10 +2,9 @@
 set -eu
 BASEDIR="$(dirname $0)"
 APP_NAME="$(echo 'const c = require("./package.json"); console.log(c.name);' | node -)"
-APP_VERSION="$(echo 'const c = require("./package.json"); console.log(c.version);' | node -)"
-echo "Running artifacts for ${APP_NAME} v${APP_VERSION}"
+export APP_VERSION="$(echo 'const c = require("./package.json"); console.log(c.version);' | node -)"
 
-. ./server/version.sh
+echo "Running artifacts for ${APP_NAME} v${APP_VERSION}"
 
 wget -q -O artifact_packaging.sh 'https://raw.githubusercontent.com/mdblp/tools/dblp/artifact/artifact_packaging.sh'
 wget -q -O artifact_images.sh 'https://raw.githubusercontent.com/mdblp/tools/dblp/artifact/artifact_images.sh'
@@ -42,10 +41,6 @@ fi
 if [ ! -d "node_modules" ]; then
   npm install
 fi
-if [ ! -d "server/node_modules" ]; then
-  cp -v .npmrc server/
-  bash -c 'cd server && npm install && npm run "security-checks"'
-fi
 
 bash build.sh
 
@@ -54,14 +49,13 @@ npm run test-cloudfront
 
 # Publish only on the main node version build
 # TODO: Get node version using: "$(node --version | cut -c 2-)" to make this script usable on another build system ?
-if [ "${ARTIFACT_NODE_VERSION}" = "${TRAVIS_NODE_VERSION:-0.0.0}" ]; then
-  mv -v dist server/dist
-  buildArchive -d "./server/" -n
-  buildDockerImage -f "server/Dockerfile" -d "server" -t "latest" -s "buildServer"
+echo ${TRAVIS_NODE_VERSION}
+if [ "10.15.3" = "${TRAVIS_NODE_VERSION:-0.0.0}" ]; then
+  buildDockerImage -f "cloudfront-dist/Dockerfile.deployment" -d . -t "latest" -r "blip-deployment" -s "final"
   publishDockerImage
   npm install --save-dev "ci-toolbox@latest"
   BUILD_SOUP="true"
   buildSOUP
 else
-  echo "Not publishing docker image: Node version ${TRAVIS_NODE_VERSION:-0.0.0} != ${ARTIFACT_NODE_VERSION}"
+  echo "Not publishing docker image: Node version ${TRAVIS_NODE_VERSION:-0.0.0}"
 fi