+++ cloudfront-dist/deployment/lib/props/WebStackProps.ts
@@ -1,0 +1,10 @@
+import * as core from '@aws-cdk/core';
+
+export interface WebStackProps extends core.StackProps {
+    domainName?: string;
+    zone?: string;
+    FrontAppName?: string;
+    version?: string;
+    prefix?: string;
+    rootBucketName?: string;
+}