+++ config/local.sh
@@ -1,25 +1,24 @@
 ## Common
-export API_HOST='http://localhost:8009'
+export API_HOST='http://localhost:3000'
 export WEBPACK_DEVTOOL='source-map'
-export PORT='3000'
+export PORT='3001'
 
-# Tidepool specific
-export I18N_ENABLED='false'
-export ALLOW_CHANGE_PASSWORD='true'
-export ALLOW_SIGNUP_PATIENT='true'
-export ALLOW_PATIENT_CHANGE_EMAIL='true'
-export ALLOW_PATIENT_CHANGE_PASSWORD='true'
-export HIDE_DONATE='false'
-export HIDE_DEXCOM_BANNER='false'
-export HIDE_UPLOAD_LINK='false'
-export CAN_SEE_PWD_LOGIN='false'
-export BRANDING='tidepool'
-export PASSWORD_MIN_LENGTH='8'
+# Diabeloop specific
+export I18N_ENABLED='true'
+export ALLOW_CHANGE_PASSWORD='false'
+export ALLOW_SIGNUP_PATIENT='false'
+export ALLOW_PATIENT_CHANGE_EMAIL='false'
+export ALLOW_PATIENT_CHANGE_PASSWORD='false'
+export HIDE_DONATE='true'
+export HIDE_DEXCOM_BANNER='true'
+export HIDE_UPLOAD_LINK='true'
+export CAN_SEE_PWD_LOGIN='true'
+export BRANDING='diabeloop'
+export PASSWORD_MIN_LENGTH='10'
 export PASSWORD_MAX_LENGTH='72'
 export MAX_FAILED_LOGIN_ATTEMPTS='5'
 export DELAY_BEFORE_NEXT_LOGIN_ATTEMPT='1'
 export LATEST_TERMS='1970-01-01'
-
 # External services:
 # URL to HELP system
 export HELP_LINK='disabled'
@@ -36,3 +35,5 @@ export SUPPORT_WEB_ADDRESS='https://example.com/'
 export REGULATORY_WEB_ADDRESS='https://example.com/'
 # Crowdin translation service (enabled / disabled):
 export CROWDIN='disabled'
+# CF config
+export TARGET_ENVIRONMENT=dev