+++ server/server.js
@@ -1,38 +1,41 @@
 /* eslint-disable lodash/prefer-lodash-typecheck */
 const http = require('http');
-const https = require('https');
 const fs = require('fs');
 const path = require('path');
 const express = require('express');
 const helmet = require('helmet');
 const bodyParser = require('body-parser');
 const compression = require('compression');
-const crypto = require('crypto');
 const morgan = require('morgan');
+const request = require('request');
 
-const jsonPackage = require('./package.json');
-const blipConfig = require('./config.app');
-const serverConfig = require('./config.server');
+let httpPort = 3000;
+
+/** @type {string[]} */
+const fileList = [];
+/** @type {http.Server} */
+let httpServer = null;
+let lambdaUrl = 'http://localhost:9001';
+
+// If ports specified override the default value (3000)
+if (process.env.PORT !== undefined && process.env.PORT !== '') {
+  httpPort = Number.parseInt(process.env.PORT, 10);
+}
+
+if (process.env.LAMBDA_URL !== undefined && process.env.LAMBDA_URL !== '') {
+  lambdaUrl = process.env.LAMBDA_URL;
+}
 
 function getStaticDir(defaultDir) {
   let dir = null;
   if (process.argv.length === 3) {
-    dir = process.argv[2];
+    dir = path.resolve(process.argv[2]);
   } else {
-    dir = path.join(__dirname, defaultDir);
+    dir = path.resolve(defaultDir);
   }
   console.info(`Serving from: '${dir}'`);
   return dir;
 }
-
-const reUrl = /(^https?:\/\/[^/]+).*/;
-/** @type {string[]} */
-const fileList = [];
-let indexHTML = '<html></html>';
-/** @type {http.Server} */
-let httpServer = null;
-/** @type {https.Server} */
-let httpsServer = null;
 
 /**
  * Get the list of files we can serve
@@ -54,48 +57,49 @@ function fetchFilesList(dir) {
  */
 function redirectMiddleware(req, res, next) {
   const reqURL = req.url;
-  const file = path.basename(reqURL);
-
-  if (file === 'index.html') {
-    // Send the modified index.html
-    res.header('Cache-Control', 'public, max-age=0');
-    res.header('Content-Type', 'text/html; charset=utf-8');
-    res.send(res.locals.htmlWithNonces);
-    return;
-  }
-
-  if (fileList.includes(file)) {
-    if (reqURL === `/${file}`) {
-      // Serve the file with 'serve-static'
-      return next();
-    }
-    // Do a redirect to the file -> can be in cache in the browser.
-    return res.redirect(`/${file}`);
-  }
-
-  // Not found, send the modified index.html by default (no 404)
-  res.header('Cache-Control', 'public, max-age=0');
-  res.header('Content-Type', 'text/html; charset=utf-8');
-  res.send(res.locals.htmlWithNonces);
-}
-
-function nonceMiddleware(req, res, next) {
-  // Set a unique nonce for each request
-  res.locals.nonce = crypto.randomBytes(16).toString('base64');
-  res.locals.htmlWithNonces = indexHTML.replace(/<(script)/g, `<$1 nonce="${res.locals.nonce}"`);
-  next();
-}
-
-function printVersion() {
-  const now = new Date().toISOString();
-  console.log(`${now} ${jsonPackage.name} v${jsonPackage.version}`);
-}
-
-function cacheIndexHTML(dir) {
-  // Cache static html file to avoid reading it from the filesystem on each request
-  const now = new Date().toISOString();
-  console.log(`${now} Caching static HTML`);
-  indexHTML = fs.readFileSync(`${dir}/index.html`, 'utf8');
+  const payload = {
+    'Records': [
+      {
+        'cf': {
+          'request': {
+            'uri': reqURL
+          }
+        }
+      }
+    ]
+  };
+  const options = {
+    method: 'POST',
+    url: `${lambdaUrl}/2015-03-31/functions/func/invocations`,
+    headers: {
+      'Content-Type': 'application/json'
+    },
+    body: JSON.stringify(payload)
+  };
+  request(options, (error, response) => {
+      if (error) {
+        console.error(error);
+        res.status(500).send('lambda middleware issue');
+        next(error);
+      } else {
+        const resBody = JSON.parse(response.body);
+        if (resBody.status !== undefined) {
+          // set headers
+          for (const hd in resBody.headers) {
+            res.header(hd, resBody.headers[hd][0].value);
+          }
+          // set body
+          if (resBody.bodyEncoding === 'base64') {
+            const b = Buffer.from(resBody.body, 'base64');
+            res.status(resBody.status).send(b);
+          } else {
+            res.status(resBody.status).send(resBody.body);
+          }
+        } else {
+          return next();
+        }
+      }
+  });
 }
 
 /**
@@ -109,120 +113,19 @@ async function stopServer(app) {
     httpServer.removeAllListeners();
     httpServer = null;
   }
-  if (httpsServer !== null) {
-    httpsServer.close();
-    httpsServer.removeAllListeners();
-    httpsServer = null;
-  }
 
   if (app !== null) {
     app.removeAllListeners();
   }
 }
 
-const contentSecurityPolicy = {
-  directives: {
-    defaultSrc: ["'self'", "'report-sample'"],
-    baseUri: ["'none'"],
-    scriptSrc: [
-      "'report-sample'",
-      "'self'",
-      "'unsafe-eval'", // At least used by pdfkit
-      (req, res) => {
-        return `'nonce-${res.locals.nonce}'`;
-      },
-    ],
-    styleSrc: [
-      "'self'",
-      'blob:',
-      "'unsafe-inline'",
-    ],
-    imgSrc: [
-      "'self'",
-      'data:',
-    ],
-    fontSrc: ["'self'", 'data:'],
-    reportUri: '/event/csp-report/violation',
-    objectSrc: ['blob:'],
-    workerSrc: ["'self'", 'blob:'],
-    childSrc: ["'self'", 'blob:'],
-    frameSrc: ["'none'"],
-    connectSrc: [
-      serverConfig.apiHost,
-    ],
-  },
-  reportOnly: false,
-};
+const staticDir = getStaticDir(`${__dirname}/../dist/static`);
 
-if (blipConfig.BRANDING === 'tidepool') {
-  console.log('Tidepool enable, setting up CSP');
-  contentSecurityPolicy.directives.scriptSrc.push('https://d12wqas9hcki3z.cloudfront.net');
-  contentSecurityPolicy.directives.scriptSrc.push('https://d33v4339jhl8k0.cloudfront.net');
-  contentSecurityPolicy.directives.styleSrc.push('https://djtflbt20bdde.cloudfront.net');
-  contentSecurityPolicy.directives.styleSrc.push('https://fonts.googleapis.com');
-  contentSecurityPolicy.directives.connectSrc.push('https://api.github.com/repos/tidepool-org/chrome-uploader/releases');
-  contentSecurityPolicy.directives.connectSrc.push('wss://*.pusher.com');
-  contentSecurityPolicy.directives.connectSrc.push('*.sumologic.com');
-  contentSecurityPolicy.directives.connectSrc.push('sentry.io');
-  contentSecurityPolicy.directives.childSrc.push('https://docs.google.com');
-  if (contentSecurityPolicy.directives.frameSrc[0] === "'none'") {
-    contentSecurityPolicy.directives.frameSrc.splice(0, 1);
-  }
-  contentSecurityPolicy.directives.frameSrc.push('https://docs.google.com');
-}
-
-if (typeof blipConfig.HELP_LINK === 'string' && blipConfig.HELP_LINK.startsWith('https://')) {
-  // Assume Zendesk
-  console.log('Zendesk enable, setting up CSP');
-  const helpUrl = blipConfig.HELP_LINK.replace(reUrl, '$1');
-  contentSecurityPolicy.directives.scriptSrc.push(helpUrl);
-  contentSecurityPolicy.directives.connectSrc.push(helpUrl);
-  contentSecurityPolicy.directives.imgSrc.push(helpUrl);
-  contentSecurityPolicy.directives.connectSrc.push('https://ekr.zdassets.com');
-  contentSecurityPolicy.directives.connectSrc.push('https://diabeloop.zendesk.com');
-}
-
-if (serverConfig.matomoUrl !== null) {
-  console.log('Matomo enable, setting up CSP');
-  const matomoUrl = serverConfig.matomoUrl.replace(reUrl, '$1');
-  contentSecurityPolicy.directives.scriptSrc.push(matomoUrl);
-  contentSecurityPolicy.directives.imgSrc.push(matomoUrl);
-  contentSecurityPolicy.directives.connectSrc.push(matomoUrl);
-}
-
-if (serverConfig.crowdinPreview) {
-  console.log('Crowdin enable, setting up CSP');
-  const crowdinURL = 'https://crowdin.com';
-  const crowdinCDN = 'https://cdn.crowdin.com/';
-  contentSecurityPolicy.directives.scriptSrc.push(crowdinCDN, crowdinURL, "'unsafe-inline'");
-  // Disable nonce because crowdin use JQuery, and it's breaking it
-  for (let i=0; i<contentSecurityPolicy.directives.scriptSrc.length; i++) {
-    if (typeof contentSecurityPolicy.directives.scriptSrc[i] === 'function') {
-      contentSecurityPolicy.directives.scriptSrc.splice(i, 1);
-      break;
-    }
-  }
-  contentSecurityPolicy.directives.imgSrc.push(crowdinCDN, 'https://crowdin-static.downloads.crowdin.com', 'https://www.gravatar.com', 'https://*.wp.com');
-  contentSecurityPolicy.directives.styleSrc.push(crowdinCDN, 'https://fonts.googleapis.com');
-  contentSecurityPolicy.directives.connectSrc.push(crowdinCDN);
-  contentSecurityPolicy.directives.fontSrc.push(crowdinCDN, 'https://fonts.gstatic.com');
-  contentSecurityPolicy.directives.objectSrc.push("'self'");
-  if (contentSecurityPolicy.directives.frameSrc[0] === "'none'") {
-    contentSecurityPolicy.directives.frameSrc.splice(0, 1);
-  }
-  contentSecurityPolicy.directives.frameSrc.push(crowdinCDN, crowdinURL, 'https://accounts.crowdin.com');
-}
-
-const staticDir = getStaticDir('dist');
-
-printVersion();
-cacheIndexHTML(staticDir);
 fetchFilesList(staticDir);
 const app = express();
 app.use(morgan(':date[iso] :remote-addr - :remote-user [:date[clf]] ":method :url HTTP/:http-version" :status :res[content-length]'));
 app.use(compression());
 app.use(helmet());
-app.use(nonceMiddleware, helmet.contentSecurityPolicy(contentSecurityPolicy));
 app.use(bodyParser.json({
   type: ['json', 'application/csp-report'],
 }));
@@ -241,35 +144,22 @@ app.use(express.static(staticDir, {
   index: false,
 }));
 
-// If no ports specified, just start on default HTTP port
-if (!(serverConfig.httpPort || serverConfig.httpsPort)) {
-  serverConfig.httpPort = 3000;
-}
-
-if (serverConfig.httpPort) {
-  httpServer = http.createServer(app).listen(serverConfig.httpPort, () => {
+if (httpPort) {
+  httpServer = http.createServer(app).listen(httpPort, () => {
     const now = new Date().toISOString();
-    console.log(`${now} Connect server started on HTTP port`, serverConfig.httpPort);
-    console.log(`${now} Serving static directory "${staticDir}/"`);
+    console.log(`${now} Connect server started on HTTP port`, httpPort);
+    console.log(`${now} Serving static directory '${staticDir}/'`);
   });
 }
 
-if (serverConfig.httpsPort && serverConfig.httpsConfig) {
-  httpsServer = https.createServer(serverConfig.httpsConfig, app).listen(serverConfig.httpsPort, () => {
-    const now = new Date().toISOString();
-    console.log(`${now} Connect server started on HTTP port`, serverConfig.httpPort);
-    console.log(`${now} Serving static directory "${staticDir}/"`);
-  });
-}
+// Handle simple process kill
+process.once('SIGTERM', async () => {
+  await stopServer(app);
+});
 
-  // Handle simple process kill
-  process.once('SIGTERM', async () => {
-    await stopServer(app);
-  });
-
-  // Handle Ctrl+C when launch in a console
-  process.once('SIGINT', async () => {
-    await stopServer(app);
-  });
+// Handle Ctrl+C when launch in a console
+process.once('SIGINT', async () => {
+  await stopServer(app);
+});
 
 module.exports = app;