+++ server/start-docker.sh
@@ -1,4 +1,0 @@
-#!/bin/sh
-set -eu
-npm run build-config
-exec node server.js